renamed analyzer module
