+++ src/network.rs
@@ -1,1 +1,2 @@
+pub mod analyzer;
 pub mod p2p;
