+++ src/network/analyzer.rs
